Move the RGL grid components into a common folder
