+++ src/ui/widgets/index.ts
@@ -3,8 +3,8 @@ import log from "loglevel";
 export { EmbeddedDisplay } from "./EmbeddedDisplay/embeddedDisplay";
 export { Label } from "./Label/label";
 export { Display } from "./Display/display";
-export { DisplayResponsive } from "./DisplayResponsive/displayResponsive";
-export { DisplayGridLayout } from "./DisplayGridLayout/displayGridLayout";
+export { DisplayResponsive } from "./DisplayReactGridLayout/displayResponsive";
+export { DisplayGridLayout } from "./DisplayReactGridLayout/displayGridLayout";
 export { Shape } from "./Shape/shape";
 export { FlexContainer } from "./FlexContainer/flexContainer";
 export { DynamicPageWidget } from "./DynamicPage/dynamicPage";
